Completed Graphing further backup fixes
Made the graphs import data and plot it. Fixed parts of backup engine
but still needs testing to make sure it is fully functional. There may
be an error in plotting time on the graphs, but I will have to wait for
real testing results to clarify. Next is to buy the smell detecting kit
and work out how to get data from it.