Added Bar Charts and started Statistics engine
Made small bug fixes. Added the ability to plot bar charts with year
group along the X. Started to make a window that calculated stats. Next
I shall complete the stats window. Add the functionality to line graphs
and make it find a statistical model as well as add the hardware! So far
the delete button on the bar chart doesn't work... no idea why...